Added Event class and use it in ResultsReader classes
The main motivation is to allow c# indexer syntax, i.e. event["_index"].
Added a set of type converters for convenience of the developers using
the SDK. Among them, only the string converter is implicit. Refer to the
type converter guidelines in MSDN. Data time converter is not added
because there's no universal format that we can use.